Create Creating a SparkSession
Creating a SparkSession
Let's start with creating a new SparkSession. In this course, you will be usually provided with one, but creating a new one or getting an existing one is a must-have skill!

Instructions

Import SparkSession from pyspark.sql.
Make a new SparkSession called "my_spark" using SparkSession.builder.getOrCreate().
Print my_spark to the console to verify it's a SparkSession.